Move register form to client component
